+++ src/liturgist/__init__.py
@@ -16,10 +16,16 @@ from .core import (
     next_sunday,
     read_schedule,
 )
+from .hymnal import (
+    load_hymnal_scores,
+    parse_hymn_number,
+)
 
 __all__ = [
     "read_schedule",
     "load_template_from_file",
     "get_scripture_text",
     "next_sunday",
+    "load_hymnal_scores",
+    "parse_hymn_number",
 ]